Added the concept of bridges
Changed the Island connections to be a Bridge with their own properties. This allows for filtering on each bridge to determine if it is a valid route to take. For example, if a bridge is a toll bridge but we do not want to take toll bridges, it can be filtered out as a valid route.